# Optimize number formatting