Moved `NumberFormatInfo.GetInstance` calls further down so that it can be skipped for more format variants. Changed byte to use uint for formatting.
I plan to move byte parsing also to uint for consistency, but that has some unexpected performance issues that I need to resolve first.

Stats for hex format (X):

|          Method |     Old | New |
|---------------- |---------:|---------:|
|   x64 TryFormatByte | 30.95 ns | 21.30 ns |
|  x64 TryFormatInt32 | 45.08 ns | 36.05 ns |
| x64 TryFormatUInt32 | 43.85 ns | 36.36 ns |
|   x86 TryFormatByte | 27.48 ns | 17.51 ns |
|  x86 TryFormatInt32 | 43.52 ns | 32.34 ns |
| x86 TryFormatUInt32 | 43.32 ns | 31.96 ns |